Header: separate ProjectHeader and CCHeader into own files
